requirements.txt
fastapi==0.110.0
uvicorn==0.30.0
webdavclient3==3.14.6
python-dotenv==1.0.1
ffmpeg-python==0.2.0
python-multipart==0.0.9
aiofiles==23.2.1 